+++ client/rpc/src/test/kotlin/net/corda/client/rpc/RPCPermissionsTests.kt
@@ -1,9 +1,6 @@
 package net.corda.client.rpc
 
-import net.corda.core.flows.FlowLogic
-import net.corda.core.messaging.CordaRPCOps
 import net.corda.core.messaging.RPCOps
-import net.corda.node.services.Permissions
 import net.corda.node.services.messaging.rpcContext
 import net.corda.nodeapi.internal.config.User
 import net.corda.testing.node.internal.RPCDriverDSL
@@ -119,6 +116,17 @@ class RPCPermissionsTests : AbstractRPCTest() {
                 proxy.validatePermission("startTrackedFlowDynamic", "net.corda.flows.OtherFlow")
             }
             proxy.validatePermission("networkMapFeed")
+        }
+    }
+
+    @Test
+    fun `killing flows requires permission`() {
+
+        rpcDriver {
+            val proxy = testProxyFor(userOf("joe", emptySet()))
+            assertNotAllowed {
+                proxy.validatePermission("killFlow")
+            }
         }
     }